% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{col_from_j}
\alias{col_from_j}
\title{Convert the given index into a column name.}
\usage{
col_from_j(j)
}
\arguments{
\item{j}{Column index}
}
\description{
Convert the given index into a column name.
}
\details{
\itemize{
\item The given column index must be an integer, i.e. \code{0L} instead of \code{0}.
\item Column names count from 1, and are strings.
}
}
\examples{
col_from_j(0L)  # returns '1'

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
