% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{i_from_row}
\alias{i_from_row}
\title{Convert the given row name into an index number.}
\usage{
i_from_row(row)
}
\description{
Convert the given row name into an index number.
}
\examples{
i_from_row('A')  # returns 0L

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
