% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{ij_from_row_col}
\alias{ij_from_row_col}
\title{Convert the given row and column names into indices.}
\usage{
ij_from_row_col(row, col)
}
\description{
Convert the given row and column names into indices.
}
\examples{
ij_from_row_col('A', '2')  # returns list(0L, 1L)

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
