% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{ij_from_well}
\alias{ij_from_well}
\title{Convert the given well name into row and column indices.}
\usage{
ij_from_well(well)
}
\description{
Convert the given well name into row and column indices.
}
\examples{
ij_from_well('A2')  # returns list(0L, 1L)

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
