% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{iter_col_indices}
\alias{iter_col_indices}
\title{Yield all of the well indices in the given column(s).}
\usage{
iter_col_indices(pattern)
}
\arguments{
\item{pattern}{A string identifying either a single column (e.g. "1") or
several, using the pattern syntax (e.g. "1,2", "1-3", "1,3,...,11").}
}
\description{
Yield all of the well indices in the given column(s).
}
\examples{
iter_col_indices('1')  # returns c(0L)
iter_col_indices('1,2')  # returns c(0L, 1L)
iter_col_indices('1-3')  # returns c(0L, 1L, 2L)

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
