% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{iter_ij_in_block}
\alias{iter_ij_in_block}
\title{Find all of the well indices in the given block.}
\usage{
iter_ij_in_block(top_left_ij, width, height)
}
\arguments{
\item{top_left_ij}{Row/column indices for the top-left corner of the block.}

\item{width}{The number of columns spanned by the block.}

\item{height}{The number of rows spanned by the block.}
}
\description{
Find all of the well indices in the given block.
}
\details{
\itemize{
\item The given indices and dimensions must be integers, i.e. \code{1L} instead of
\code{1}.
}
}
\examples{
iter_ij_in_block(list(0L, 1L), 2L, 2L)

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
