% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{j_from_col}
\alias{j_from_col}
\title{Convert the given column name into an index number.}
\usage{
j_from_col(col)
}
\description{
Convert the given column name into an index number.
}
\details{
Note that column names count from 1.
}
\examples{
j_from_col('1')  # returns 0L

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
