% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{row_col_from_ij}
\alias{row_col_from_ij}
\title{Convert the given indices into row and column names.}
\usage{
row_col_from_ij(i, j)
}
\arguments{
\item{i}{Row index}

\item{j}{Column index}
}
\description{
Convert the given indices into row and column names.
}
\details{
The given indices must be integers, i.e. \code{0L} instead of \code{0}.
}
\examples{
row_col_from_ij(0L, 1L)  # returns list('A', '2')

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
