% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{row_col_from_well}
\alias{row_col_from_well}
\title{Split row and column names out of the given well name.}
\usage{
row_col_from_well(well)
}
\arguments{
\item{well}{Well name}
}
\description{
Split row and column names out of the given well name.
}
\details{
The well name is allowed to be zero-padded.
}
\examples{
row_col_from_well('A2')  # returns list('A', '2')

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
