% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{row_from_i}
\alias{row_from_i}
\title{Convert the given index into a row name.}
\usage{
row_from_i(i)
}
\arguments{
\item{i}{Row index}
}
\description{
Convert the given index into a row name.
}
\details{
\itemize{
\item The given row index must be an integer, i.e. \code{0L} instead of \code{0}.
\item The row after 'Z' is 'AA'.
}
}
\examples{
row_from_i(0L)  # returns 'A'

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
