% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{show}
\alias{show}
\title{Visualize the given microplate layout.}
\usage{
show(toml_path, params = NULL, style = NULL)
}
\arguments{
\item{toml_path}{The path to a file describing the layout of one or more plates.  See the
\href{https://wellmap.readthedocs.io/en/latest/file_format.html}{file format documentation}
for details about this file.}

\item{params}{The names of one or more experimental parameters from the above TOML file to
visualize.  For example, if the TOML file contains something equivalent to
\code{well.A1.conc = 1}, then \code{"conc"} would be a valid parameter name.  If not
specified, the default is to display any parameters that have at least two
different values.}

\item{style}{Settings that control miscellaneous aspects of the plot, e.g. colors,
dimensions, etc.  See \link{Style} for more info.}
}
\description{
It's wise to visualize TOML layouts before doing any analysis, to ensure
that all of the wells are correctly annotated.  The \code{wellmap} command-line
program is a useful tool for doing this, but sometimes it's more convenient
to make visualizations directly from R/Rstudio.  That's what this function
is for.
}
