% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{show_df}
\alias{show_df}
\title{Visualize the microplate layout described by the given data frame.}
\usage{
show_df(df, cols = NULL, style = NULL)
}
\arguments{
\item{df}{The data frame describing the layout to plot.  The data frame must be
\href{http://vita.had.co.nz/papers/tidy-data.html}{tidy}: each row must describe
a single well, and each column must describe a single aspect of each well.
The location of each well must be specified using one or more of the same
columns that wellmap uses for that purpose, namely:
\itemize{
\item \code{plate}
\item \code{path}
\item \code{well}
\item \code{well0}
\item \code{row}
\item \code{col}
\item \code{row_i}
\item \code{col_j}
}

See \code{\link[=load]{load()}} for the exact meanings of these columns.  It's not necessary to
specify all of these columns, there just needs to be enough information to
locate each well.  If the \emph{plate} column is missing, it is assumed that all
of the wells are on the same plate.  It is also assumed that any redundant
columns (e.g. \emph{row} and \emph{row_i}) will be consistent with each other.

Any scalar-valued columns other than these can be plotted.}

\item{cols}{Which columns to plot onto the layout.  The columns used to locate the
wells (listed above) cannot be plotted.  The default is to include any
columns that have at least two different values.}

\item{style}{Settings than control miscellaneous aspects of the plot, e.g. colors,
dimensions, etc.  See \link{Style} for more info.}
}
\description{
Unlike the \code{\link[=show]{show()}} function and the \code{wellmap} command-line program, this
function is not limited to displaying layouts parsed directly from TOML
files.  Any data frame that specifies a well for each row can be plotted.
This provides the means to:
\itemize{
\item Project experimental data onto a layout.
\item Visualize layouts that weren't generated by wellmap in the first place.
}

For example, you could load experimental data into a data frame and use
this function to visualize it directly, without ever having to specify a
layout.  This might be a useful way to get a quick sense for the data.
}
