% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{Style}
\alias{Style}
\title{Describe how to plot well layouts.}
\usage{
Style(
  .kw_only = NULL,
  cell_size = 0.25,
  pad_width = 0.2,
  pad_height = 0.2,
  bar_width = 0.15,
  bar_pad_width = 0.2,
  top_margin = 0.5,
  left_margin = 0.5,
  right_margin = 0.2,
  bottom_margin = 0.2,
  color_scheme = "rainbow"
)
}
\arguments{
\item{.kw_only}{This just exists to check that all the other arguments are named, and not
just specified positionally.  The order of these arguments is not guaranteed
to remain the same in different versions of wellmap.}

\item{cell_size}{The size of the boxes representing each well, in inches.  The default is
0.25".}

\item{pad_width}{The vertical padding between layouts, in inches.  The default is 0.2".}

\item{pad_height}{The horizontal padding between layouts, in inches.  The default is 0.2".}

\item{bar_width}{The width of the color bar, in inches.  The default is 0.15".}

\item{bar_pad_width}{The horizontal padding between the color bar and the nearest layout, in
inches.  The default is 0.2".}

\item{top_margin}{The space between the layouts and the top edge of the figure, in inches.
The default is 0.5".}

\item{left_margin}{The space between the layouts and the left edge of the figure, in inches.
The default is 0.5".}

\item{right_margin}{The space between the layouts and the right edge of the figure, in inches.
The default is 0.2".}

\item{bottom_margin}{The space between the layouts and the bottom edge of the figure, in inches.
The default is 0.2".}

\item{color_scheme}{The name of the color scheme to use.  Each different value for each
different parameter will be assigned a color from this scheme.  Any name
understood by either \href{http://colorcet.pyviz.org/}{colorcet} or
\href{https://matplotlib.org/examples/color/colormaps_reference.html}{matplotlib}
can be used.}
}
\description{
Style objects exist to be passed to \code{\link[=show]{show()}} or \code{\link[=show_df]{show_df()}}, where they
determine various aspects of the plots' appearances.  The following
parameters can be specified:
}
