% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{well0_from_row_col}
\alias{well0_from_row_col}
\title{Create a zero-padded well name from the given row and column names.}
\usage{
well0_from_row_col(row, col, digits = 2L)
}
\description{
Create a zero-padded well name from the given row and column names.
}
\details{
If specified, the number of column digits must be an integer, i.e. \code{2L}
instead of \code{2}.
}
\examples{
well0_from_row_col('A', '2')  # returns 'A02'

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
