% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{well0_from_well}
\alias{well0_from_well}
\title{Create a zero-padded well name from the given well name.}
\usage{
well0_from_well(well, digits = 2L)
}
\description{
Create a zero-padded well name from the given well name.
}
\details{
\itemize{
\item It doesn't matter if the input name is zero-padded or not.
\item If specified, the number of column digits must be an integer, i.e. \code{2L}
instead of \code{2}.
}
}
\examples{
well0_from_well('A2')  # returns 'A02'

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
