% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{well_from_ij}
\alias{well_from_ij}
\title{Create a well name from the given row and column indices.}
\usage{
well_from_ij(i, j)
}
\arguments{
\item{i}{Row index}

\item{j}{Column index}
}
\description{
Create a well name from the given row and column indices.
}
\details{
The given indices must be integers, i.e. \code{0L} instead of \code{0}.
}
\examples{
well_from_ij(0L, 1L)  # returns 'A2'

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
