% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{well_from_row_col}
\alias{well_from_row_col}
\title{Create a well name from the given row and column names.}
\usage{
well_from_row_col(row, col)
}
\description{
Create a well name from the given row and column names.
}
\examples{
well_from_row_col('A', '2')  # returns 'A2'

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
