% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{iter_row_indices}
\alias{iter_row_indices}
\title{Yield all of the well indices in the given row(s).}
\usage{
iter_row_indices(pattern)
}
\arguments{
\item{pattern}{A string identifying either a single row (e.g. "A") or
several, using the pattern syntax (e.g. "A,B", "A-C", "A,C,...,G").}
}
\description{
Yield all of the well indices in the given row(s).
}
\examples{
iter_row_indices('A')  # returns c(0L)
iter_row_indices('A,B')  # returns c(0, 1)
iter_row_indices('A-C')  # returns c(0, 1, 2)

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
