% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellmapr.R
\name{iter_well_indices}
\alias{iter_well_indices}
\title{Yield the indices for the given well(s).}
\usage{
iter_well_indices(pattern)
}
\description{
The given pattern can either specify a single well (e.g. "A1") or several,
using the pattern syntax (e.g. "A1,B2", "A1-B2", "A1,A3,...,B11").
}
\examples{
iter_well_indices('A1'))
iter_well_indices('A1,B2'))
iter_well_indices('A1-B2'))

}
\seealso{
\href{https://wellmap.readthedocs.io/en/latest/well_formats.html}{Well Formats}
}
